Don't use faulty vram measurement for deciding on (soft) worker restart
In Linux, gpu_info.get_total_vram_mb() can be 0.
If this happens, don't use this faulty measurement to decide to restart
the worker.

This tweak allows me to enable lora's on Linux. Without it, the worker
gets in a soft restart loop after generating a single lora, and then
crashes.

Please note; I have tested this to run for a few hours without crashing,
but it seems that using Lora's on linux is still leaking ram, so this is
not a full fix.